rename utils and add getters.py
