+++ module/action_potential_functions.py
@@ -12,7 +12,7 @@ from scipy.ndimage import gaussian_filter1d
 import scipy.signal as sg
 from scipy import stats
 from collections import namedtuple
-from module.base_utils import *
+from module.utils import *
 from scipy.optimize import curve_fit
 import sys
